+++ src/app/md-menu/md-menu.module.ts
@@ -1,7 +1,7 @@
 import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 import { RouterModule } from '@angular/router';
-import { TourMdMenuModule } from 'ngx-tour-md-menu';
+import { TourMatMenuModule } from 'ngx-tour-md-menu';
 
 import { DocsComponent } from './docs.component';
 import { MdMenuComponent } from './md-menu.component';
@@ -11,7 +11,7 @@ import { OtherRouteComponent } from './other-route.component';
   declarations: [MdMenuComponent, DocsComponent, OtherRouteComponent],
   imports: [
     CommonModule,
-    TourMdMenuModule.forRoot(),
+    TourMatMenuModule.forRoot(),
     RouterModule.forChild([{
       component: MdMenuComponent,
       path: '',